pyproject.toml
[project]
name = "agent-orchestrator"
version = "0.1.0"
description = "Agent Orchestrated Dev System — Agent-CLI Mode (Hexagonal + Event-Driven)"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "redis>=5.0",
    "rq>=1.16",
    "pyyaml>=6.0",
    "gitpython>=3.1",
    "click>=8.1",
    "structlog>=24.0",
    "prometheus-client>=0.20",
    "python-ulid>=2.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "fakeredis>=2.20",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
